Create seller model (#18)
Creating this pull request for the seller login view model. This should
probably throw an error since in the create-seller branch, there were
missing docs in both of view and view model file. I added just the "///"
in my view model page but didn't add anything in the view to avoid merge
conflicts.

---------

Co-authored-by: Levi Lesches <levilesches@gmail.com>